# Improve performance and reduce allocations in TemplateBinder ctor