@JamesNK 2018-11-12
> Do you only ever intend to use the first parameter transformer?

Yes, for simplicity.

> Also, why not pass the constraints and transformer(s) separately to the TemplateBinder ctor? Right now, we do the work of extracting them from the list of parameter policies twice.

It is a public constructor. If additional usages of policies in the future then the constructor would need to change.